refactor(sip): extract Registrar admission + RegisterBeeper UAC dialog FSM
Second extraction. Registrar owns the runtime mode (Open/Learn/Secure),
digest challenge/verify, and the Learn-mode TOFU + MAC-lock device
registry with its NVS persistence; RegisterBeeper owns the register-beep
outbound dialog machine (INVITE→ACK→BYE / CANCEL-on-timeout).
RequestsHandler keeps its public API via type aliases and thin wrappers.

The adopted-device online flag now lives in the Registrar's records
instead of only in the dashboard snapshot; the snapshot becomes a plain
mirror refreshed through a consumeDevicesChanged() dirty flag at the
end of each handled packet. Same visible behavior, one owner.

deserializeBlob and the pbxcfg namespace move to PbxPersist.hpp, shared
by both translation units. PbxEnv gains localIp()/serverPort().
