+++ main/CMakeLists.txt
@@ -77,6 +77,8 @@ idf_component_register(SRCS ${MAIN_SRC}
                             "../src/Helpers/OtaUpdater.cpp"
                             "../src/SIP/RequestsHandler.cpp"
                             "../src/SIP/TransactionLayer.cpp"
+                            "../src/SIP/Registrar.cpp"
+                            "../src/SIP/RegisterBeeper.cpp"
                             "../src/SIP/Session.cpp"
                             "../src/SIP/SipClient.cpp"
                             "../src/SIP/SipMessage.cpp"
