+++ src/SIP/RequestsHandler.hpp
@@ -33,6 +33,8 @@
 #include "RtpSender.hpp"
 #include "PbxEnv.hpp"
 #include "TransactionLayer.hpp"
+#include "Registrar.hpp"
+#include "RegisterBeeper.hpp"
 
 class RequestsHandler : private PbxEnv
 {
@@ -158,36 +160,20 @@ public:
 
 	// ── Registrar mode (STAGE 2) ──────────────────────────────────────────────────
 	// Runtime registrar policy, replacing the compile-time POCKETDIAL_OPEN_REGISTRAR
-	// gate. NVS-persisted (namespace "pbxcfg", key "reg_mode"); the compile-time
-	// symbol now only seeds the DEFAULT at boot. Thread-safe (mirror setDnd): setter
-	// takes _mutex + refreshes the snapshot; getter reads an atomic so the SIP hot
-	// path (onRegister) never locks just to branch on the mode.
-	enum class RegistrarMode : uint8_t
-	{
-		Open   = 0,   // standalone: accept every REGISTER, no challenge (legacy)
-		Learn  = 1,   // TOFU + MAC-lock: adopt unknown devices, enforce secured ones
-		Secure = 2,   // require digest auth for every provisioned extension
-	};
+	// gate. The policy machine itself lives in Registrar (see Registrar.hpp); these
+	// aliases + wrappers keep the public API stable for the dashboard/TUI. Setter
+	// takes _mutex; getter reads an atomic so the SIP hot path never locks.
+	using RegistrarMode = Registrar::Mode;
 	void setRegistrarMode(RegistrarMode mode);
 	RegistrarMode getRegistrarMode() const;
 
 	// ── Device registry (STAGE 2: Learn-mode adoption) ────────────────────────────
-	// Adopted-device lifecycle for the TUI. A device is keyed by its 12-hex MAC and
-	// remembers the extension it registered as and whether it has been promoted from
-	// first-seen (Learned) to digest-enforced (Secured). All accessors are
-	// thread-safe (snapshot mutex) and NVS-persisted, mirroring forwards/groups.
-	enum class DeviceState : uint8_t
-	{
-		Learned = 0,   // TOFU: seen + accepted, not yet locked to digest auth
-		Secured = 1,   // promoted: MAC-locked + digest-enforced for its extension
-	};
-	struct AdoptedDevice
-	{
-		std::string mac;         // 12 lowercase hex chars
-		std::string extension;   // the AOR it last registered as
-		DeviceState state = DeviceState::Learned;
-		bool online = false;     // currently has a live registration binding
-	};
+	// Adopted-device lifecycle for the TUI, owned by the Registrar machine. A device
+	// is keyed by its 12-hex MAC and remembers the extension it registered as and
+	// whether it has been promoted from first-seen (Learned) to digest-enforced
+	// (Secured). All accessors are thread-safe (snapshot mutex) and NVS-persisted.
+	using DeviceState = Registrar::DeviceState;
+	using AdoptedDevice = Registrar::AdoptedDevice;
 	// Snapshot of all adopted devices for the dashboard/TUI (thread-safe).
 	std::vector<AdoptedDevice> getAdoptedDevices();
 	// Promote a device to Secured (MAC-locked + digest-enforced). Accepts either a
@@ -269,44 +255,9 @@ private:
 	// Called from the single-threaded SIP handler path — no additional mutex needed.
 	void onDtmfInfo(std::shared_ptr<SipMessage> data);
 
-	// ── Register beep (signaling-only intercom tone) ─────────────────────────────
-	// On a NEW registration, send the registering phone a brief auto-answer INVITE so
-	// it plays its own intercom tone (the "beep"), then tear the call straight back
-	// down. NO RTP is ever sourced — the tone is the phone's local intercom alert. The
-	// outbound UAC dialog is tracked in a small bounded ring (_beepDialogs) keyed by
-	// Call-ID so onOk() can drive ACK→BYE and tick() can time it out / CANCEL it. All
-	// of these assume the caller already holds _mutex (non-recursive).
-	//
-	// State machine (per beep dialog):
-	//   sendRegisterBeep() : allocate a slot, send INVITE (auto-answer headers), arm
-	//                        a deadline; if no slot free, skip the beep (it's cosmetic).
-	//   onOk() INVITE 200  : send ACK, then BYE, advance to AwaitingByeOk.
-	//   onOk() BYE 200     : free the slot.
-	//   tick() deadline    : if still AwaitingInviteOk, CANCEL and free; if
-	//                        AwaitingByeOk, just free (best-effort BYE already sent).
-	void sendRegisterBeep(const std::shared_ptr<SipClient>& phone);
-	std::shared_ptr<SipMessage> buildBeepAck(const std::shared_ptr<SipMessage>& ok);
-	std::shared_ptr<SipMessage> buildBeepBye(const std::shared_ptr<SipMessage>& ok);
-	std::shared_ptr<SipMessage> buildBeepCancel(std::size_t slot);
-
-	// AwaitingCancelDone: CANCEL sent, lingering until the 487 final response is ACKed
-	// (or a bounded deadline frees the slot). Added for #90 — see beep teardown notes.
-	enum class BeepState { Free, AwaitingInviteOk, AwaitingByeOk, AwaitingCancelDone };
-	struct BeepDialog
-	{
-		BeepState state = BeepState::Free;
-		std::string callID;        // fresh per beep; how onOk()/tick() find this slot
-		std::string branch;        // Via branch (reused for INVITE/CANCEL)
-		std::string fromTag;       // our (server) From tag
-		std::string ext;           // target extension (phone number)
-		sockaddr_in addr{};        // phone's contact address
-		std::chrono::steady_clock::time_point deadline{};
-	};
-	// Bounded outbound-UAC dialog table. Tiny fixed footprint; if all slots are busy a
-	// new registration just skips its beep. Guarded by _mutex.
-	std::array<BeepDialog, POCKETDIAL_MAX_BEEPS> _beepDialogs;
-	// Find the beep slot owning a Call-ID, or nullptr. Caller holds _mutex.
-	BeepDialog* findBeepByCallID(std::string_view callID);
+	// Register beep (signaling-only intercom tone): the outbound UAC dialog FSM
+	// lives in RegisterBeeper (see RegisterBeeper.hpp). Guarded by _mutex.
+	RegisterBeeper _beeper{*this};
 
 	// ── PbxEnv: shared-infrastructure surface for the extracted machines ───────
 	// RequestsHandler is the PbxEnv implementation each decomposed state machine
@@ -324,9 +275,21 @@ private:
 	{
 		queueLog(std::move(msg), isError);
 	}
+	const std::string& localIp() const override { return _localIp; }
+	int serverPort() const override { return _serverPort; }
 
 	// RFC 3261 §17 INVITE client transactions (Timer A/B/L). Guarded by _mutex.
 	TransactionLayer _txLayer{*this};
+
+	// REGISTER admission policy + adopted-device registry (STAGE 2). Guarded by
+	// _mutex except the lock-free mode atomic. The compile-time
+	// POCKETDIAL_OPEN_REGISTRAR symbol only seeds the DEFAULT mode at boot; the
+	// NVS-persisted value (loaded in the constructor) overrides it.
+#ifdef POCKETDIAL_OPEN_REGISTRAR
+	Registrar _registrar{*this, Registrar::Mode::Open};
+#else
+	Registrar _registrar{*this, Registrar::Mode::Secure};
+#endif
 
 	// RFC 4028 session timer helpers. Caller holds _mutex.
 	void armSessionTimer(Session* session, const std::shared_ptr<SipMessage>& ok200);
@@ -580,19 +543,6 @@ private:
 	// TUI can read without taking _mutex. Persisted to NVS ("pbxcfg"/"rewarm_min").
 	std::atomic<uint16_t> _rewarmMinutes{60};
 
-	// ── Registrar mode (STAGE 2) ──────────────────────────────────────────────────
-	// Atomic so onRegister() can read the policy without taking _mutex (it already
-	// holds _mutex via handle(), but keeping this atomic also lets getRegistrarMode()
-	// be lock-free for the dashboard). Seeded from POCKETDIAL_OPEN_REGISTRAR at boot,
-	// then overridden by the persisted NVS value if present.
-#ifdef POCKETDIAL_OPEN_REGISTRAR
-	std::atomic<RegistrarMode> _registrarMode{RegistrarMode::Open};
-#else
-	std::atomic<RegistrarMode> _registrarMode{RegistrarMode::Secure};
-#endif
-	void loadRegistrarMode();             // boot-time reload from NVS; caller holds _mutex
-	void persistRegistrarMode();          // write-through after a mode change; holds _mutex
-
 	// ── Admin HTTP-open TTL (PLAN_ADMIN_HTTP_ONLY.md Phase 2) ─────────────────────
 	// How long a DTMF trigger keeps the HTTP admin plane reachable, in seconds.
 	// Atomic (mirrors _registrarMode) so HttpServer's accept-loop can read it
@@ -603,40 +553,9 @@ private:
 	std::atomic<uint16_t> _adminHttpTtlSec{600};
 	void loadAdminHttpTtl();              // boot-time reload from NVS; caller holds _mutex
 
-	// ── Device registry (STAGE 2: Learn-mode adoption) ────────────────────────────
-	// Adopted devices keyed by 12-hex MAC. Bounded by POCKETDIAL_MAX_CLIENTS (a flood
-	// of distinct MACs cannot grow the heap without limit; new keys past the cap are
-	// dropped, mirroring _dnd/_forwards). Guarded by _mutex; mirrored into the
-	// snapshot and persisted to NVS (namespace "pbxcfg", key "devices").
-	struct DeviceRecord
-	{
-		std::string extension;
-		DeviceState state = DeviceState::Learned;
-	};
-	std::unordered_map<std::string, DeviceRecord> _devices;   // mac -> record
-	void loadDevices();                   // boot-time reload; caller holds _mutex
-	void persistDevices();                // write-through after a mutation; holds _mutex
-	void refreshDeviceSnapshot();         // mirror _devices into _snapshot; holds both mutexes
-
-	// Learn-mode REGISTER admission. Caller holds _mutex. Resolves the source MAC,
-	// applies TOFU + MAC-lock, and returns the digest decision. On a first-packet
-	// ARP miss it returns Accept (deferring the lock to the next REGISTER). May set
-	// `outRejectReason` when returning Reject. Records/updates the adoption entry.
-	enum class AuthDecision { Accept, Challenge, Reject };
-	AuthDecision admitLearn(const std::shared_ptr<SipMessage>& data,
-		const std::string& ext, std::string& outRejectReason);
-	// Secure-mode REGISTER admission: challenge + verify digest for `ext`. Caller
-	// holds _mutex. Enqueues the 401 (with WWW-Authenticate) itself when challenging.
-	AuthDecision admitSecure(const std::shared_ptr<SipMessage>& data,
-		const std::string& ext, std::string& outRejectReason);
-	// Emit a 401 Unauthorized with a fresh WWW-Authenticate challenge for `data`
-	// into _outbox. `stale` answers an expired-but-valid nonce. Caller holds _mutex.
-	void sendChallenge(const std::shared_ptr<SipMessage>& data, bool stale);
-	// Emit a 403 Forbidden with a reason phrase into _outbox. Caller holds _mutex.
-	void sendForbidden(const std::shared_ptr<SipMessage>& data, const std::string& reason);
-	// Mark a device record online/offline in the snapshot after a (de)registration.
-	// Caller holds _mutex.
-	void markDeviceOnline(const std::string& mac, bool online);
+	// Mirror the Registrar's adopted-device registry into the dashboard snapshot.
+	// Caller holds _mutex; takes _snapshotMutex internally.
+	void refreshDeviceSnapshot();
 
 	// NVS persistence for _forwards / _ringGroups / _pageZones. No-ops on host (the
 	// maps are the store); on ESP they read/write the "pbxcfg" NVS namespace.
